Factor timeout utils to their own module.
